Explicity cast to double for clarity
Add a cast to make clear integer conversion is intentional.

+++ app/src/main/java/de/markusfisch/android/shadereditor/hardware/CameraListener.java
@@ -238,7 +238,8 @@ public class CameraListener {
 		Camera.Size bestSizeAspect = null;
 
 		for (Camera.Size size : sizes) {
-			double diff = Math.abs(size.height - height) +
+			double diff = (double)
+					Math.abs(size.height - height) +
 					Math.abs(size.width - width);
 
 			if (diff < minDiff) {
